requirements.txt
# Core dependencies
fastapi>=0.115.2
uvicorn[standard]==0.30.0
python-multipart>=0.0.18
pydantic==2.11.7
pydantic-settings==2.5.0
starlette>=0.40.0

# DepthFlow
depthflow==0.9.1
torch  # 如果有 GPU，可改為 torch+cu118 或適合的版本

# Async task queue
celery==5.3.6
redis==5.2.0

# Image processing
pillow==10.4.0
numpy>=1.26.0

# File handling
aiofiles==24.1.0

# Development
python-dotenv==1.0.1

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0

# Monitoring
prometheus-client==0.21.0

# CORS (built into FastAPI, no separate package needed)